+++ client/src/App.js
@@ -7,6 +7,7 @@ import Register from './components/auth/Register';
 import Login from './components/auth/Login';
 import Alert from './components/layout/Alert';
 import Dashboard from './components/dashboard/Dashboard';
+import CreateProfile from './components/profile-forms/CreateProfile';
 
 import PrivateRoute from './components/routing/PrivateRoute';
 
@@ -44,6 +45,9 @@ const App = () => {
                         </Route>
                         <PrivateRoute exact path='/dashboard'>
                             <Dashboard />
+                        </PrivateRoute>
+                        <PrivateRoute exact path='/create-profile'>
+                            <CreateProfile />
                         </PrivateRoute>
                     </Switch>
                 </section>